requirements.txt
fastapi
uvicorn
httpx
ollama
python-dotenv
openai
google-generativeai

